refactor: lib 디렉토리 구조 개선 - 대형 파일 분할 및 기능별 정리
## 주요 변경사항

### 🏗️ 새로운 디렉토리 구조
- `lib/services/miso/` - MISO 관련 서비스 로직 분리
- `lib/config/` - 설정 파일들 기능별 분할

### 📁 대형 파일 분할
- `miso-knowledge.ts` (577줄) → 3개 파일로 분할
  - `services/miso/parser.ts` - 파싱 로직 (147줄)
  - `services/miso/api.ts` - API 호출 (148줄)
  - `services/miso/workflow.ts` - 워크플로우 (264줄)

- `constants/config.ts` (461줄) → 5개 파일로 분할
  - `config/upload.ts` - 파일 업로드 설정
  - `config/workflow.ts` - 워크플로우 설정
  - `config/ui.ts` - UI 관련 설정
  - `config/api.ts` - API 및 환경 설정
  - `config/defaults.ts` - 기본값 프리셋

### 🔧 코드 품질 개선
- 중복 함수 제거 (`formatFileSize` 통합)
- 기능별 책임 분리 (파싱, API, 워크플로우)
- import 경로 정리

### 🔄 레거시 호환성
- 기존 import 경로 유지 (re-export 방식)
- 점진적 마이그레이션 지원
- 기존 API 변경 없음
