+++ lib/services/miso/index.ts
@@ -1,0 +1,23 @@
+/**
+ * MISO Services 통합 익스포트
+ */
+
+// Parser utilities
+export { 
+  parseInterviewDetail, 
+  groupInterviewsByTopic,
+  type InterviewTopicData 
+} from './parser'
+
+// API client functions
+export { 
+  createMisoDocumentOnly, 
+  checkDocumentStatus, 
+  addSegmentsToDocument 
+} from './api'
+
+// Workflow management
+export { 
+  syncInterviewTopicsToPersona, 
+  getOrCreateTopicId 
+} from './workflow'
